Refactor buttons to separate binary sensor component
