+++ esphome/components/tt21100/__init__.py
@@ -1,0 +1,5 @@
+import esphome.codegen as cg
+
+CODEOWNERS = ["@kroimon"]
+
+tt21100_ns = cg.esphome_ns.namespace("tt21100")
