+++ esphome/components/tt21100/touchscreen/tt21100.cpp
@@ -1,11 +1,48 @@
 #include "tt21100.h"
-#include "esphome/core/helpers.h"
 #include "esphome/core/log.h"
 
 namespace esphome {
 namespace tt21100 {
 
 static const char *const TAG = "tt21100";
+
+static const uint8_t MAX_BUTTONS = 4;
+static const uint8_t MAX_TOUCH_POINTS = 5;
+static const uint8_t MAX_DATA_LEN = (7 + MAX_TOUCH_POINTS * 10);  // 7 Header + (Points * 10 data bytes)
+
+struct TT21100ButtonReport {
+  uint16_t length;     // Always 14 (0x000E)
+  uint8_t report_id;   // Always 0x03
+  uint16_t timestamp;  // Number in units of 100 us
+  uint8_t btn_value;   // Only use bit 0..3
+  uint16_t btn_signal[MAX_BUTTONS];
+} __attribute__((packed));
+
+struct TT21100TouchRecord {
+  uint8_t : 5;
+  uint8_t touch_type : 3;
+  uint8_t tip : 1;
+  uint8_t event_id : 2;
+  uint8_t touch_id : 5;
+  uint16_t x;
+  uint16_t y;
+  uint8_t pressure;
+  uint16_t major_axis_length;
+  uint8_t orientation;
+} __attribute__((packed));
+
+struct TT21100TouchReport {
+  uint16_t length;
+  uint8_t report_id;
+  uint16_t timestamp;
+  uint8_t : 2;
+  uint8_t large_object : 1;
+  uint8_t record_num : 5;
+  uint8_t report_counter : 2;
+  uint8_t : 3;
+  uint8_t noise_effect : 3;
+  TT21100TouchRecord touch_record[0];
+} __attribute__((packed));
 
 void TT21100TouchscreenStore::gpio_intr(TT21100TouchscreenStore *store) { store->touch = true; }
 
@@ -54,14 +91,13 @@ void TT21100Touchscreen::loop() {
       // Button event
       auto *report = (TT21100ButtonReport *) data;
 
-      ESP_LOGV(TAG, "Button report: Len=%d, ID=%d, Time=%5u, Val=[%u] - [%04X][%04X][%04X][%04X]", report->length,
-               report->report_id, report->timestamp, report->btn_val, report->btn_signal[0], report->btn_signal[1],
-               report->btn_signal[2], report->btn_signal[3]);
+      ESP_LOGV(TAG, "Button report: Len=%d, ID=%d, Time=%5u, Value=[%u], Signal=[%04X][%04X][%04X][%04X]",
+               report->length, report->report_id, report->timestamp, report->btn_value, report->btn_signal[0],
+               report->btn_signal[1], report->btn_signal[2], report->btn_signal[3]);
 
-      for (int i = 0; i < 4; i++) {
-        if (this->buttons_[i] != nullptr) {
-          this->buttons_[i]->publish_state(report->btn_signal[i] > 0);
-        }
+      for (uint8_t i = 0; i < 4; i++) {
+        for (auto *listener : this->button_listeners_)
+          listener->update_button(i, report->btn_signal[i]);
       }
 
     } else if (data_len >= 7) {
@@ -69,9 +105,9 @@ void TT21100Touchscreen::loop() {
       auto *report = (TT21100TouchReport *) data;
 
       ESP_LOGV(TAG,
-               "Touch report: Len=%d, ID=%d, Time=%5u, LargeObject=%u, RecordNum=%u, RecordCounter=%u, NoiseEfect=%u",
+               "Touch report: Len=%d, ID=%d, Time=%5u, LargeObject=%u, RecordNum=%u, RecordCounter=%u, NoiseEffect=%u",
                report->length, report->report_id, report->timestamp, report->large_object, report->record_num,
-               report->report_counter, report->noise_efect);
+               report->report_counter, report->noise_effect);
 
       uint8_t touch_count = (data_len - sizeof(TT21100TouchReport)) / sizeof(TT21100TouchRecord);
 
@@ -133,10 +169,6 @@ void TT21100Touchscreen::dump_config() {
   LOG_I2C_DEVICE(this);
   LOG_PIN("  Interrupt Pin: ", this->interrupt_pin_);
   LOG_PIN("  Reset Pin: ", this->reset_pin_);
-  LOG_BINARY_SENSOR("  ", "Button 1", this->buttons_[0]);
-  LOG_BINARY_SENSOR("  ", "Button 2", this->buttons_[1]);
-  LOG_BINARY_SENSOR("  ", "Button 3", this->buttons_[2]);
-  LOG_BINARY_SENSOR("  ", "Button 4", this->buttons_[3]);
 }
 
 }  // namespace tt21100
